Add credentials polling
